Refactor tests, 100% coverage
