+++ src/index.ts
@@ -1,7 +1,7 @@
-export { WindowPostMessageStream } from './window/WindowPostMessageStream';
-export { WorkerPostMessageStream } from './WebWorker/WorkerPostMessageStream';
-export { WorkerParentPostMessageStream } from './WebWorker/WorkerParentPostMessageStream';
-export { ParentProcessMessageStream } from './node-process/ParentProcessMessageStream';
-export { ChildProcessMessageStream } from './node-process/ChildProcessMessageStream';
+export * from './window/WindowPostMessageStream';
+export * from './WebWorker/WorkerPostMessageStream';
+export * from './WebWorker/WorkerParentPostMessageStream';
+export * from './node-process/ParentProcessMessageStream';
+export * from './node-process/ChildProcessMessageStream';
 export { PostMessageEvent } from './BasePostMessageStream';
 export { StreamData, StreamMessage } from './utils';
